refactor(searchInput): add props to repare storybook
