+++ src/ui/components/inputs/types.ts
@@ -25,6 +25,8 @@ type CustomSearchInputProps = {
   searchInputID?: string
   onFocusState?: (focus: boolean) => void
   showLocationButton?: boolean
+  onPressLocationButton?: () => void
+  locationLabel?: string
 }
 
 export type RNTextInputProps = Pick<
@@ -86,6 +88,9 @@ export function getCustomSearchInputProps(props: SearchInputProps): CustomSearch
     accessibilityDescribedBy: props.accessibilityDescribedBy,
     onPressRightIcon: props.onPressRightIcon,
     focusOutlineColor: props.focusOutlineColor,
+    showLocationButton: props.showLocationButton,
+    onPressLocationButton: props.onPressLocationButton,
+    locationLabel: props.locationLabel,
   }
 }
